FYP2026/requirement.txt
streamlit
pandas
faker
langchain
langchain-community
sqlalchemy
langchain-google-genai
google-generativeai